+++ src/lsm/table.zig
@@ -292,7 +292,7 @@ pub fn TableType(
             }
 
             const DataFinishOptions = struct {
-                cluster: u32,
+                cluster: u128,
                 address: u64,
                 snapshot_min: u64,
                 tree_id: u16,
@@ -389,7 +389,7 @@ pub fn TableType(
             }
 
             const IndexFinishOptions = struct {
-                cluster: u32,
+                cluster: u128,
                 address: u64,
                 snapshot_min: u64,
                 tree_id: u16,
